refactor: extract hooks, utilities, and components from profile pages
Edit page:
- useProfileAuth: auth guard hook
- useEditProfileData: data fetching and form hydration hook
- useProfileSubmit: full submit pipeline hook (avatar, upserts, polling, session)
- useProfileSelectOptions: memoized multi-select option lists
- EditPageHeader: back/cancel/save button bar component
- parseUserExperiences: pure parsing functions (parseLinks, parseWhatIOffer, parseWorkExperiences, parseEducations)
- pollUntilSynced: polling utility extracted from onSubmit

Profile page:
- platformLabelMap: moved to shared social-links module
- ProfileBadgeSection: reusable badge-list component replacing 5 repeated blocks
- scheduleFormatters: formatSelectedDate, formatStartTimeSlot, toDateKey helpers
